requirements.txt
crispy-bootstrap5==0.7
Django==4.2.2
django-crispy-forms==2.0
django-jazzmin==2.6.0
django-money==3.1.0
django-shortuuidfield==0.1.3
Pillow==9.5.0
psycopg==3.2.2
psycopg2-binary==2.9.9
py-moneyed==2.0
shortuuid==1.0.11

